+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Check for `settings.json` version vs. DynAIkonTrap version in case settings are copied from one trap to another
+
+### Fixed
+- Implementation of UrSense interface following updated documentation
+
+### Changed
+- Video sensor logs to JSON for easier machine reading -- parsing this back to the previous VTT output is trivial
+- Interface to initialise `Output` -- output mode is now handled internally
 
 ---